On create, values need to be read from the RIL, not set
This causes problem on dual SIM phones where setting one SIM to
3G would cause the RIL to automatically change the other RIL to
2G.

In this case, when the setting saved is 3G for both SIM, whenever
the user switches tab, the tab the user switches to become 3G.
For example, when the user switches to Sim 2, Sim 2 becomes 3G.
And when the user switches back to Sim 1, Sim 1 becomes 3G. This
can be very confusing, not to mention that switching between
tabs should not cause the network mode to change.

The updateBody() call is removed to ensure that when SET is called,
the value is not immediately GET from the RIL as the RIL might not
finish updating the value yet and GET will return the old value,
resulting in the following behavior:

- User is on 2G
- User switches to 3G
- UI still displays "2G" due to updating value from RIL
- After a few seconds, UI displays "3G" because RIL update is
  finished.

Not sure what the call is for, but it's probably not used in GSM
phones. Care should be taken if this patch is to be applied to
other phones.

From: Pawit Pornkitprasan <p.pawit@gmail.com>